Install Active Admin
I have opted for it instead of RailsAdmin because of my experiences on
https://gitlab.com/pragmacode/media_choice. There we started with
RailsAdmin and have rewritten the admin using Active Admin. The main
problems with RailsAdmin were:

* harder to customize
* making all models manageable by the admin led to a polluted admin
interface where the final user had trouble to find the correct options

Regarding this installation it is the result of following the
instructions at: https://activeadmin.info/0-installation.html.