Update rails and postgres-copy gems
We had to update postgres-copy as part of rails' update because the previous version required rails<5.1.

The modified configuration files are the minimum necessary to get the tests to pass and the csv download to work.
There will be another commit with all the configuration files updates.

The rails upgrade is necessary so we can use yarn to manage javascript dependencies.

Signed-off-by: Rafael Reggiani Manzo <rr.manzo@protonmail.com>
Signed-off-by: Gabriely Pereira <gabriely.pereira@usp.br>
Signed-off-by: yurickyh <yurickyussuke@hotmail.com>
Signed-off-by: Rodolfo Alves <rodolfosalves@lsdi.ufma.br>